Add files via upload
Fine-tuning DistilBERT on FINREP dataset.
This is the first part of the process which consist of removing certain stop-words, the newline character, parenthesis and the text present within parenthesis from the column 'Text'. Duplicate entries and text with word count less than 7 are removed. The training, validation and test data consist of the 'Topic' column and the preprocessed 'Text' column. The split data is saved as data_clean.